move db basics to ra_db
This should allow to move hir to a separate crate
